pyproject.toml
[tool.poetry]
name = "client-activity-monitor"
version = "0.1.0"
description = "Monitor user activity across multiple Oracle databases"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
oracledb = {version = "^2.0.0", extras = ["kerberos"]}
pydantic = "^2.0.0"
customtkinter = "^5.2.0"
pandas = "^2.2.0"
openpyxl = "^3.1.2"
pyyaml = "^6.0"
loguru = "^0.7.0"
pyperclip = "^1.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
pylint = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
client-monitor = "client_activity_monitor.main:main"